+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk2/Configuration.kt
@@ -358,7 +358,7 @@ class Configuration(var context: Context,
             }
 
             if (simulcastEnabled) {
-                it.enableSimulcast(simulcastRid)
+                //it.enableSimulcast(simulcastRid)
             }
         }
     }